LayoutTests/fast/xpath/nsresolver-bad-object-expected.txt

CONSOLE MESSAGE: TypeError: 'lookupNamespaceURI' property of callback interface should be callable
SUCCESS
